Make Maven comparisons fair again
We were measuring the time it takes Gradle to clean the output folder
while Maven didn't do that work (as it has no concept of incremental
builds). We now don't measure the clean time in any case.

We were comparing `gradle assemble` (which does no tests) with
`mvn package` (which runs unit tests). We now explicitly exclude
tests for Maven.